refactor: :recycle: カスタムフックの作成
